fixing ui bug with draft documents and the toast that tells you that drafts can't be annotated
